+++ packages/core/src/storages/memoryStorage.ts
@@ -13,7 +13,7 @@ import { createStorage } from '../createStorage.ts'
  * import { defineConfig } from 'kubb'
  *
  * export default defineConfig({
- *   input:  { path: './petStore.yaml' },
+ *   input:  './petStore.yaml',
  *   output: { path: './src/gen' },
  *   storage: memoryStorage(),
  * })